move methods to model
